modificações entidades criação dos verbos put e post com validações

+++ src/main/java/br/com/gustavokonzen/api_escriba/service/CartorioService.java
@@ -1,6 +1,10 @@
 package br.com.gustavokonzen.api_escriba.service;
 
 import br.com.gustavokonzen.api_escriba.dto.CartorioDTO;
+import br.com.gustavokonzen.api_escriba.exception.CampoNaoEncontradoException;
+import br.com.gustavokonzen.api_escriba.exception.ErroNaoMapeadoException;
+import br.com.gustavokonzen.api_escriba.exception.RegistroDuplicadoException;
+import br.com.gustavokonzen.api_escriba.exception.RegistroNaoEncontradoException;
 import br.com.gustavokonzen.api_escriba.model.Cartorio;
 import br.com.gustavokonzen.api_escriba.repository.CartorioRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -8,8 +12,8 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -29,32 +33,50 @@ public class CartorioService {
 
     @Transactional
     public Cartorio salvar(Cartorio cartorio) {
+        if(Objects.isNull(cartorio)){
+            throw new ErroNaoMapeadoException("Cartório não pode ser vazio");
+        }
+        if (cartorio.getNome().isEmpty()) {
+            throw new CampoNaoEncontradoException("O campo 'Nome' é obrigatório");
+        }
+        var cartorioDB = cartorioRepository.findByNome(cartorio.getNome());
+        if (Objects.nonNull(cartorioDB) && cartorio.getNome().equals(cartorioDB.getNome())) {
+            throw new RegistroDuplicadoException("Nome já informado no cartório: " + cartorioDB.getId());
+        }
         return cartorioRepository.save(cartorio);
     }
 
     @Transactional
     public Cartorio atualizar(Integer id, Cartorio cartorioAtualizado) {
-        Optional<Cartorio> cartorioOptional = cartorioRepository.findById(id);
+        var cartorioDB = cartorioRepository.findByNome(cartorioAtualizado.getNome());
 
-        Cartorio cartorioExistente = cartorioOptional.get();
-
+        if(Objects.isNull(cartorioDB)){
+            cartorioDB = cartorioRepository.findById(id).get();
+        }
+        if(cartorioDB.getId() != id && cartorioDB.getNome().equals(cartorioAtualizado.getNome())){
+            throw new RegistroDuplicadoException("Nome já informado no cartório: " + cartorioDB.getId());
+        }
         if (cartorioAtualizado.getNome() != null) {
-            cartorioExistente.setNome(cartorioAtualizado.getNome());
+            cartorioDB.setNome(cartorioAtualizado.getNome());
         }
         if (cartorioAtualizado.getObservacao() != null) {
-            cartorioExistente.setObservacao(cartorioAtualizado.getObservacao());
+            cartorioDB.setObservacao(cartorioAtualizado.getObservacao());
         }
         if (cartorioAtualizado.getSituacao() != null) {
-            cartorioExistente.setSituacao(cartorioAtualizado.getSituacao());
+            cartorioDB.setSituacao(cartorioAtualizado.getSituacao());
         }
         if (cartorioAtualizado.getAtribuicaoCartorioList() != null && !cartorioAtualizado.getAtribuicaoCartorioList().isEmpty()) {
-            cartorioExistente.setAtribuicaoCartorioList(cartorioAtualizado.getAtribuicaoCartorioList());
+            cartorioDB.setAtribuicaoCartorioList(cartorioAtualizado.getAtribuicaoCartorioList());
         }
 
-        return cartorioRepository.save(cartorioExistente);
+        return cartorioRepository.save(cartorioDB);
     }
 
     public void deletar(int id) {
-        cartorioRepository.deleteById(id);
+        if (cartorioRepository.findById(id).isPresent()) {
+            cartorioRepository.deleteById(id);
+        } else {
+            throw new RegistroNaoEncontradoException("Cartório: " + id + " não encontrado!");
+        }
     }
 }
